Improve rename/move dialog, it now takes a free text path and detects the library to which it belongs by itself

+++ core/gamedex-core-game/src/main/kotlin/com/gitlab/ykrasik/gamedex/core/game/presenter/DeleteGamePresenter.kt
@@ -52,7 +52,12 @@ class DeleteGamePresenter @Inject constructor(
 
         private suspend fun onAccept() {
             if (view.fromFileSystem.value) {
-                fileSystemService.delete(game.path)
+                if (game.path.exists()) {
+                    fileSystemService.delete(game.path)
+                } else {
+                    view.onError(game.path.toString(), title = "File doesn't exist!")
+                    return
+                }
             }
 
             taskService.execute(gameService.delete(game))
